#876 reconcile_related_queries First steps.
